requirements.txt
pandas 
numpy
pymongo 
llama-cpp-python
python-dotenv
langchain
langchain-community
google-generativeai 
streamlit